organized the examples better
