+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/MainActivity.java
@@ -20,31 +20,31 @@ import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.github.javiersantos.materialstyleddialogs.MaterialStyledDialog;
 import com.mapbox.mapboxandroiddemo.adapter.ExampleAdapter;
-import com.mapbox.mapboxandroiddemo.examples.AnimateMapCameraActivity;
-import com.mapbox.mapboxandroiddemo.examples.AnimatedMarkerActivity;
-import com.mapbox.mapboxandroiddemo.examples.BasicMarkerViewActivity;
-import com.mapbox.mapboxandroiddemo.examples.BasicUserLocation;
-import com.mapbox.mapboxandroiddemo.examples.BoundingBoxCameraActivity;
-import com.mapbox.mapboxandroiddemo.examples.CustomInfoWindowActivity;
-import com.mapbox.mapboxandroiddemo.examples.CustomRasterStyleActivity;
-import com.mapbox.mapboxandroiddemo.examples.DefaultStyleActivity;
-import com.mapbox.mapboxandroiddemo.examples.DirectionsActivity;
-import com.mapbox.mapboxandroiddemo.examples.DrawCustomMarkerActivity;
-import com.mapbox.mapboxandroiddemo.examples.DrawGeojsonLineActivity;
-import com.mapbox.mapboxandroiddemo.examples.DrawMarkerActivity;
-import com.mapbox.mapboxandroiddemo.examples.DrawPolygonActivity;
-import com.mapbox.mapboxandroiddemo.examples.GeocodingActivity;
-import com.mapbox.mapboxandroiddemo.examples.LocationTrackingActivity;
-import com.mapbox.mapboxandroiddemo.examples.MapMatchingActivity;
-import com.mapbox.mapboxandroiddemo.examples.MapboxMapOptionActivity;
-import com.mapbox.mapboxandroiddemo.examples.MapboxStudioStyleActivity;
-import com.mapbox.mapboxandroiddemo.examples.OfflineManagerActivity;
-import com.mapbox.mapboxandroiddemo.examples.SatelliteStyleActivity;
-import com.mapbox.mapboxandroiddemo.examples.SimpleMapViewActivity;
-import com.mapbox.mapboxandroiddemo.examples.SimpleOfflineMapActivity;
-import com.mapbox.mapboxandroiddemo.examples.SimplifyPolylineActivity;
-import com.mapbox.mapboxandroiddemo.examples.StaticImageActivity;
-import com.mapbox.mapboxandroiddemo.examples.SupportMapFragmentActivity;
+import com.mapbox.mapboxandroiddemo.examples.camera.AnimateMapCameraActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.AnimatedMarkerActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.BasicMarkerViewActivity;
+import com.mapbox.mapboxandroiddemo.examples.location.BasicUserLocation;
+import com.mapbox.mapboxandroiddemo.examples.camera.BoundingBoxCameraActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.CustomInfoWindowActivity;
+import com.mapbox.mapboxandroiddemo.examples.styles.CustomRasterStyleActivity;
+import com.mapbox.mapboxandroiddemo.examples.styles.DefaultStyleActivity;
+import com.mapbox.mapboxandroiddemo.examples.mas.DirectionsActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.DrawCustomMarkerActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.DrawGeojsonLineActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.DrawMarkerActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.DrawPolygonActivity;
+import com.mapbox.mapboxandroiddemo.examples.mas.GeocodingActivity;
+import com.mapbox.mapboxandroiddemo.examples.location.LocationTrackingActivity;
+import com.mapbox.mapboxandroiddemo.examples.mas.MapMatchingActivity;
+import com.mapbox.mapboxandroiddemo.examples.basics.MapboxMapOptionActivity;
+import com.mapbox.mapboxandroiddemo.examples.styles.MapboxStudioStyleActivity;
+import com.mapbox.mapboxandroiddemo.examples.offline.OfflineManagerActivity;
+import com.mapbox.mapboxandroiddemo.examples.styles.SatelliteStyleActivity;
+import com.mapbox.mapboxandroiddemo.examples.basics.SimpleMapViewActivity;
+import com.mapbox.mapboxandroiddemo.examples.offline.SimpleOfflineMapActivity;
+import com.mapbox.mapboxandroiddemo.examples.mas.SimplifyPolylineActivity;
+import com.mapbox.mapboxandroiddemo.examples.mas.StaticImageActivity;
+import com.mapbox.mapboxandroiddemo.examples.annotations.SupportMapFragmentActivity;
 import com.mapbox.mapboxandroiddemo.labs.LocationPickerActivity;
 import com.mapbox.mapboxandroiddemo.labs.MarkerFollowingRouteActivity;
 import com.mapbox.mapboxandroiddemo.labs.SpaceStationLocationActivity;
