+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/examples/camera/AnimateMapCameraActivity.java
@@ -1,25 +1,25 @@
-package com.mapbox.mapboxandroiddemo.examples;
+package com.mapbox.mapboxandroiddemo.examples.camera;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 
 import com.mapbox.mapboxandroiddemo.R;
+import com.mapbox.mapboxsdk.camera.CameraPosition;
 import com.mapbox.mapboxsdk.camera.CameraUpdateFactory;
 import com.mapbox.mapboxsdk.geometry.LatLng;
-import com.mapbox.mapboxsdk.geometry.LatLngBounds;
 import com.mapbox.mapboxsdk.maps.MapView;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
 
-public class BoundingBoxCameraActivity extends AppCompatActivity {
+public class AnimateMapCameraActivity extends AppCompatActivity {
 
     private MapView mapView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_camera_bounding_box);
+        setContentView(R.layout.activity_camera_animate);
 
         mapView = (MapView) findViewById(R.id.mapView);
         mapView.onCreate(savedInstanceState);
@@ -27,17 +27,19 @@ public class BoundingBoxCameraActivity extends AppCompatActivity {
             @Override
             public void onMapReady(final MapboxMap mapboxMap) {
 
-                // When user clicks the map, fit the camera to the bounding box
+                // When user clicks the map, animate to new camera location
                 mapboxMap.setOnMapClickListener(new MapboxMap.OnMapClickListener() {
                     @Override
                     public void onMapClick(@NonNull LatLng point) {
-                        LatLngBounds latLngBounds = new LatLngBounds.Builder()
-                                .include(new LatLng(36.532128, -93.489121)) // Northeast
-                                .include(new LatLng(25.837058, -106.646234)) // Southwest
-                                .build();
+                        CameraPosition position = new CameraPosition.Builder()
+                                .target(new LatLng(51.50550, -0.07520)) // Sets the new camera position
+                                .zoom(17) // Sets the zoom
+                                .bearing(180) // Rotate the camera
+                                .tilt(30) // Set the camera tilt
+                                .build(); // Creates a CameraPosition from the builder
 
-                        mapboxMap.easeCamera(CameraUpdateFactory.newLatLngBounds(latLngBounds, 50), 5000);
-
+                        mapboxMap.animateCamera(CameraUpdateFactory
+                                .newCameraPosition(position), 7000);
                     }
                 });
             }
