+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/examples/mas/MapMatchingActivity.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxandroiddemo.examples;
+package com.mapbox.mapboxandroiddemo.examples.mas;
 
 import android.graphics.Color;
 import android.os.AsyncTask;
@@ -8,13 +8,18 @@ import android.text.TextUtils;
 import android.util.Log;
 
 import com.mapbox.mapboxandroiddemo.R;
+import com.mapbox.mapboxsdk.MapboxAccountManager;
 import com.mapbox.mapboxsdk.annotations.PolylineOptions;
 import com.mapbox.mapboxsdk.geometry.LatLng;
 import com.mapbox.mapboxsdk.maps.MapView;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
+import com.mapbox.services.commons.ServicesException;
+import com.mapbox.services.commons.geojson.LineString;
 import com.mapbox.services.commons.models.Position;
-import com.mapbox.services.commons.utils.PolylineUtils;
+import com.mapbox.services.directions.v4.DirectionsCriteria;
+import com.mapbox.services.mapmatching.v4.MapboxMapMatching;
+import com.mapbox.services.mapmatching.v4.models.MapMatchingResponse;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -26,9 +31,13 @@ import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SimplifyPolylineActivity extends AppCompatActivity {
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
-    private static final String TAG = "SimplifyLineActivity";
+public class MapMatchingActivity extends AppCompatActivity {
+
+    private static final String TAG = "MapMatchingActivity";
 
     private MapView mapView;
     private MapboxMap map;
@@ -36,13 +45,14 @@ public class SimplifyPolylineActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_simplify_polyline);
+        setContentView(R.layout.activity_mas_map_matching);
 
-        mapView = (MapView) findViewById(R.id.mapview);
+        mapView = (MapView) findViewById(R.id.mapView);
         mapView.onCreate(savedInstanceState);
         mapView.getMapAsync(new OnMapReadyCallback() {
             @Override
             public void onMapReady(MapboxMap mapboxMap) {
+
                 map = mapboxMap;
 
                 new DrawGeoJSON().execute();
@@ -89,7 +99,7 @@ public class SimplifyPolylineActivity extends AppCompatActivity {
 
             try {
                 // Load GeoJSON file
-                InputStream inputStream = getAssets().open("matched_route.geojson");
+                InputStream inputStream = getAssets().open("trace.geojson");
                 BufferedReader rd = new BufferedReader(new InputStreamReader(inputStream, Charset.forName("UTF-8")));
                 StringBuilder sb = new StringBuilder();
                 int cp;
@@ -130,13 +140,16 @@ public class SimplifyPolylineActivity extends AppCompatActivity {
         protected void onPostExecute(List<Position> points) {
             super.onPostExecute(points);
 
-            drawBeforeSimplify(points);
-            drawSimplify(points);
+            drawBeforeMapMatching(points);
+
+            // Convert the route to a linestring
+            LineString lineString = LineString.fromCoordinates(points);
+            drawMapMatched(lineString, 8);
 
         }
     }
 
-    private void drawBeforeSimplify(List<Position> points) {
+    private void drawBeforeMapMatching(List<Position> points) {
 
         LatLng[] pointsArray = new LatLng[points.size()];
         for (int i = 0; i < points.size(); i++)
@@ -148,21 +161,52 @@ public class SimplifyPolylineActivity extends AppCompatActivity {
                 .width(4));
     }
 
-    private void drawSimplify(List<Position> points) {
+    private void drawMapMatched(LineString lineString, int precision) {
+        try {
+            // Setup the request using a client.
+            MapboxMapMatching client = new MapboxMapMatching.Builder()
+                    .setAccessToken(MapboxAccountManager.getInstance().getAccessToken())
+                    .setProfile(DirectionsCriteria.PROFILE_DRIVING)
+                    .setGpsPrecison(precision)
+                    .setTrace(lineString)
+                    .build();
 
-        Position[] before = new Position[points.size()];
-        for (int i = 0; i < points.size(); i++) before[i] = points.get(i);
+            // Execute the API call and handle the response.
+            client.enqueueCall(new Callback<MapMatchingResponse>() {
+                @Override
+                public void onResponse(Call<MapMatchingResponse> call, Response<MapMatchingResponse> response) {
 
-        Position[] after = PolylineUtils.simplify(before, 0.001);
+                    // Create a new list to store the map matched coordinates.
+                    List<LatLng> mapMatchedPoints = new ArrayList<>();
 
-        LatLng[] result = new LatLng[after.length];
-        for (int i = 0; i < after.length; i++)
-            result[i] = new LatLng(after[i].getLatitude(), after[i].getLongitude());
+                    // Check that the map matching API response is "OK".
+                    if (response.code() == 200) {
+                        // Convert the map matched response list from position to latlng coordinates.
+                        for (int i = 0; i < response.body().getMatchedPoints().length; i++) {
+                            mapMatchedPoints.add(new LatLng(
+                                    response.body().getMatchedPoints()[i].getLatitude(),
+                                    response.body().getMatchedPoints()[i].getLongitude()));
+                        }
 
-        map.addPolyline(new PolylineOptions()
-                .add(result)
-                .color(Color.parseColor("#3bb2d0"))
-                .width(4));
+                        // Add the map matched route to the Mapbox map.
+                        map.addPolyline(new PolylineOptions()
+                                .addAll(mapMatchedPoints)
+                                .color(Color.parseColor("#3bb2d0"))
+                                .width(4));
+                    } else {
+                        // If the response code does not response "OK" an error has occurred.
+                        Log.e(TAG, "Too many coordinates, Profile not found, invalid input, or no match");
+                    }
+                }
 
+                @Override
+                public void onFailure(Call<MapMatchingResponse> call, Throwable t) {
+                    Log.e(TAG, "MapboxMapMatching error: " + t.getMessage());
+                }
+            });
+        } catch (ServicesException e) {
+            Log.e(TAG, "MapboxMapMatching error: " + e.getMessage());
+            e.printStackTrace();
+        }
     }
 }
