+++ MapboxAndroidDemo/src/main/res/layout/activity_annotation_animated_marker.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.AnimatedMarkerActivity">
+    tools:context=".examples.annotations.AnimatedMarkerActivity">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
