+++ MapboxAndroidDemo/src/main/res/layout/activity_annotation_custom_info_window.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.SimpleMapViewActivity">
+    tools:context=".examples.basics.SimpleMapViewActivity">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
