+++ MapboxAndroidDemo/src/main/res/layout/activity_basic_simple_mapview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.SimpleMapViewActivity">
+    tools:context=".examples.basics.SimpleMapViewActivity">
 
     <!-- Set the starting camera position and map style using xml-->
     <com.mapbox.mapboxsdk.maps.MapView
